# docs/openapi.yaml
openapi: 3.0.0
info:
  title: DealPig API
  description: |
    API documentation for the DealPig application.
    This API provides endpoints for managing leads, campaigns, email tracking, and more.
  version: 1.0.0
  contact:
    email: support@dealpig.com
servers:
  - url: https://app.dealpig.com/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        code:
          type: string
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
...
paths:
...
  /tracking:
    get:
      summary: Track email opens via pixel
      description: Records when an email is opened by loading a tracking pixel
      operationId: trackEmailOpen
      parameters:
        - name: id
          in: query
          description: Email tracking ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 1x1 transparent pixel image
          content:
            image/png:
              schema:
                type: string
                format: binary
        '400':
          description: Missing tracking ID
          
    post:
      summary: Track email events
      description: Records email events like bounces and replies
      operationId: trackEmailEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - event
                - trackingId
              properties:
                event:
                  type: string
                  enum: [bounce, reply]
                trackingId:
                  type: string
                reason:
                  type: string
      responses:
        '200':
          description: Event recorded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'